+++ src/main/java/com/project/view/AdminView.java
@@ -1,0 +1,10 @@
+package main.java.com.project.view;
+
+import main.java.com.project.dto.Member;
+
+public class AdminView {
+    public void run(Member member){
+        System.out.println("adminView");
+        System.out.println(member.getEmail());
+    }
+}
